refactor: SwiftData Model 폴더로 관리
